requirements_dev.txt
flask
gevent
celery
gitpython